requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
azure-cognitiveservices-speech==1.40.0
cartesia==1.0.14
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
daily-python==0.13.0
distro==1.9.0
fastapi==0.115.5
flatbuffers==24.3.25
frozenlist==1.5.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
humanfriendly==10.0
idna==3.10
iniconfig==2.0.0
iterators==0.2.0
jiter==0.7.1
llvmlite==0.43.0
loguru==0.7.2
Markdown==3.7
mpmath==1.3.0
multidict==6.1.0
numba==0.60.0
numpy==1.26.4
onnxruntime==1.19.2
openai==1.50.2
packaging==24.2
pillow==10.4.0
pipecat-ai==0.0.49
pluggy==1.5.0
propcache==0.2.0
protobuf==4.25.5
pydantic==2.8.2
pydantic_core==2.20.1
pyloudnorm==0.1.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-deepcompare==1.0.1
python-dotenv==1.0.1
requests==2.32.3
resampy==0.4.3
scipy==1.14.1
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.3
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==13.1
yarl==1.17.2